requirements.txt
PyQt6>=6.6.0
pytest>=7.4.0
pytest-qt>=4.2.0
python-dateutil>=2.8.0

